+++ src/global_functions.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 use Symfony\Component\Process\ExecutableFinder;
 use Symfony\Component\String\Slugger\AsciiSlugger;
 use Tools\Exceptionist;
+use Tools\Filesystem;
 use function Symfony\Component\String\u;
 
 if (!defined('IS_WIN')) {
@@ -94,12 +95,13 @@ if (!function_exists('is_json')) {
      * Checks if a string is JSON
      * @param string $string String
      * @return bool
+     * @deprecated 1.7.4 Use instead `json_validate()`
      */
     function is_json(string $string): bool
     {
-        json_decode($string);
+        deprecationWarning('Deprecated. Use instead `json_validate()`');
 
-        return json_last_error() === JSON_ERROR_NONE;
+        return json_validate($string);
     }
 }
 
@@ -119,8 +121,7 @@ if (!function_exists('is_stringable')) {
     /**
      * Checks is a value can be converted to string.
      *
-     * Arrays that can be converted to strings with `array_to_string ()` are
-     *  stringable.
+     * Arrays that can be converted to strings with `array_to_string ()` are stringable.
      * @param mixed $var A var you want to check
      * @return bool
      * @since 1.2.5
@@ -142,18 +143,34 @@ if (!function_exists('is_stringable')) {
 if (!function_exists('objects_map')) {
     /**
      * Executes an object method for all objects of the given arrays
-     * @param array<object> $objects An array of objects. Each object must have
-     *  the method to be called
+     * @param array<object> $objects An array of objects. Each object must have the method to be called
      * @param string $method The method to be called for each object
      * @param array $args Optional arguments for the method to be called
-     * @return array Returns an array containing all the returned values of the
-     *  called method applied to each object
+     * @return array Returns an array containing all the returned values of the called method applied to each object
      * @throws \Tools\Exception\MethodNotExistsException
+     * @deprecated 1.7.4 deprecated, it will be removed in a later release
      * @since 1.1.11
      */
     function objects_map(array $objects, string $method, array $args = []): array
     {
-        return array_map(fn (object $object) => call_user_func_array(Exceptionist::methodExists($object, $method), $args), $objects);
+        deprecationWarning('`objects_map()` is deprecated and will be removed in a later release');
+
+        return array_map(fn(object $object) => call_user_func_array(Exceptionist::methodExists($object, $method), $args), $objects);
+    }
+}
+
+if (!function_exists('rtr')) {
+    /**
+     * Fast and convenient alias for `Filesystem::rtr()`
+     * @param string $path Absolute path
+     * @return string Relative path
+     * @throws \ErrorException
+     * @see \Tools\Filesystem::rtr()
+     * @since 1.7.4
+     */
+    function rtr(string $path): string
+    {
+        return Filesystem::rtr($path);
     }
 }